pyproject.toml
[tool.poetry]
name = "backend"
version = "0.0.1"
description = ""
authors = ["TEAM X"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.109.1"
uvicorn = {extras = ["standard"], version = "^0.27.0.post1"}
sqlalchemy = "^2.0"
asyncpg= "^0.28.0"
alembic = "^1.12.0"
python-dotenv = "^1.0.0"
pydantic-settings = "^2.1.0"
pydantic = {extras = ["email"], version = "^2.6.0"}
python-jose = "^3.3.0"
psycopg2 = "^2.9.9"
python-multipart = "^0.0.7"
bcrypt = "4.0.1"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}





[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
